LayoutTests/plugins/non-private-browsing-mode.html

<head>
<script src="../resources/js-test.js"></script>
</head>
<body onload="runTest()">
<script>
function runTest() 
{
    if (!window.testRunner) {
        debug("This test can only run from within DumpRenderTree because it requires TestNetscapePlugin.\n");
        return;
    }
    
    plugin = document.createElement("embed");
    plugin.type = "application/x-webkit-test-netscape";
    plugin.setAttribute('test', 'private-browsing');
    document.body.appendChild(plugin);
    
    shouldBe("plugin.privateBrowsingEnabled", "false");
    shouldBe("plugin.cachedPrivateBrowsingEnabled", "false");
}
</script>

<script>
description("Tests that the private browsing API implementation works as expected");


</script>
